// backend/src/common/enums/notification-type.enum.ts
export enum NotificationType {
  RESERVATION_CONFIRMED = 'RESERVATION_CONFIRMED',
  CANCEL_TOKEN_REMINDER = 'CANCEL_TOKEN_REMINDER',
  DRIVER_ASSIGNED = 'DRIVER_ASSIGNED',
  REMINDER_J1 = 'REMINDER_J1',
  REMINDER_H1 = 'REMINDER_H1',
  RIDE_STARTED = 'RIDE_STARTED',
  RIDE_COMPLETED = 'RIDE_COMPLETED',
  RESERVATION_CANCELLED = 'RESERVATION_CANCELLED',
  DRIVER_NEW_RIDE = 'DRIVER_NEW_RIDE',
  DRIVER_RIDE_MODIFIED = 'DRIVER_RIDE_MODIFIED',
  DRIVER_RIDE_CANCELLED = 'DRIVER_RIDE_CANCELLED',
  DRIVER_REMINDER_J1 = 'DRIVER_REMINDER_J1',
  ADMIN_NEW_RESERVATION = 'ADMIN_NEW_RESERVATION',
  ADMIN_DRIVER_ASSIGNED = 'ADMIN_DRIVER_ASSIGNED',
  ADMIN_UNPAID_RIDE = 'ADMIN_UNPAID_RIDE',
  ADMIN_ARCHIVE = 'ADMIN_ARCHIVE',
  MONTHLY_DRIVER_REPORT = 'MONTHLY_DRIVER_REPORT',
  MONTHLY_ADMIN_REPORT = 'MONTHLY_ADMIN_REPORT',
}